use enum names for material uniforms

+++ CrossMod/Rendering/RAnimation.cs
@@ -55,7 +55,7 @@ namespace CrossMod.Rendering
                     {
                         if (System.Enum.TryParse(a.AttributeName, out SSBHLib.Formats.Materials.MatlEnums.ParamId paramId))
                         {
-                            m.Material.MaterialAnimation.Add((long)paramId, a.Keys.GetValue(frame));
+                            m.Material.MaterialAnimation.Add(paramId, a.Keys.GetValue(frame));
                         }
                     }
                 }
